Refactor http on agent to be more like layer.
