+++ mirrord/agent/src/main.rs
@@ -6,6 +6,8 @@
 #![feature(type_alias_impl_trait)]
 #![feature(unix_socket_abstract)]
 #![feature(tcp_quickack)]
+#![feature(async_fn_in_trait)]
+#![allow(incomplete_features)]
 
 use std::{
     collections::{HashMap, HashSet},
