+++ mirrord/agent/src/steal/http/hyper_handler.rs
@@ -16,15 +16,13 @@ use tokio::{
     net::TcpStream,
     sync::{mpsc::Sender, oneshot},
 };
+use tracing::error;
 
-use super::error::HttpTrafficError;
+use super::{error::HttpTrafficError, HttpVersionT};
 use crate::{
     steal::{HandlerHttpRequest, MatchedHttpRequest},
     util::ClientId,
 };
-
-pub(super) mod httpv1;
-pub(super) mod httpv2;
 
 /// Used to pass data to the [`Service`] implementation of [`hyper`].
 ///
@@ -99,7 +97,7 @@ fn header_matches(
         })
 }
 
-async fn prepare_response(
+pub(super) async fn prepare_response(
     (mut parts, body): (response::Parts, Incoming),
 ) -> Result<Response<Full<Bytes>>, HttpTrafficError> {
     // Remove headers that would be invalid due to us fiddling with the `body`.
@@ -134,4 +132,111 @@ async fn matched_request(
     parts.headers.remove(http::header::TRANSFER_ENCODING);
 
     Ok(Response::from_parts(parts, body))
+}
+
+impl<V> HyperHandler<V>
+where
+    V: HttpVersionT + Debug,
+{
+    /// Handles the case when no filter matches a header in the request (or we have an HTTP upgrade
+    /// request).
+    ///
+    /// # HTTP/1
+    ///
+    /// ## Flow
+    ///
+    /// 1. Creates a [`http1::Connection`](hyper::client::conn::http1::Connection) to the
+    /// `original_destination`;
+    ///
+    /// 2. Sends the [`Request`] to it, and awaits a [`Response`];
+    ///
+    /// 3. Sends the [`HttpResponse`] back on the connected [`TcpStream`];
+    ///
+    /// ### Special case (HTTP upgrade request)
+    ///
+    /// If the [`Request`] is an HTTP upgrade request, then we send the `original_destination`
+    /// connection, through `upgrade_tx`, to be handled in [`filter_task`](super::filter_task).
+    ///
+    /// - Why not use [`hyper::upgrade::on`]?
+    ///
+    /// [`hyper::upgrade::on`] requires the original [`Request`] as a parameter, due to it having
+    /// the [`OnUpgrade`](hyper::upgrade::OnUpgrade) receiver tucked inside as an
+    /// [`Extensions`](http::extensions::Extensions)
+    /// (you can see this [here](https://docs.rs/hyper/1.0.0-rc.2/src/hyper/upgrade.rs.html#73)).
+    ///
+    /// [`hyper::upgrade::on`] polls this receiver to identify if this an HTTP upgrade request.
+    ///
+    /// The issue for us is that we need to send the [`Request`] to its original destination, with
+    /// [`SendRequest`](hyper::client::conn::http1::SendRequest), which takes ownership of the
+    /// request, prohibiting us to also pass it to the proper hyper upgrade handler.
+    ///
+    /// Trying to copy the [`Request`] is futile, as we can't copy the `OnUpgrade` extension, and if
+    /// we move it from the original `Request` to a copied `Request`, the channel will never
+    /// receive anything due it being in a different `Request` than the one we actually send to
+    /// the hyper machine.
+    ///
+    /// # HTTP/2
+    ///
+    /// ## Flow
+    ///
+    /// 1. Creates a [`http2::Connection`](hyper::client::conn::http2::Connection) to the
+    /// `original_destination`;
+    ///
+    /// 2. Sends the [`Request`] to it, and awaits a [`Response`];
+    ///
+    /// 3. Sends the [`HttpResponse`] back on the connected [`TcpStream`];
+    #[tracing::instrument(level = "trace")]
+    async fn unmatched_request(
+        request: Request<Incoming>,
+        upgrade_tx: Option<oneshot::Sender<RawHyperConnection>>,
+        original_destination: SocketAddr,
+    ) -> Result<Response<Full<Bytes>>, HttpTrafficError> {
+        // TODO(alex): We need a "retry" mechanism here for the client handling part, when the
+        // server closes a connection, the client could still be wanting to send a request,
+        // so we need to re-connect and send.
+        let tcp_stream = TcpStream::connect(original_destination)
+            .await
+            .inspect_err(|fail| {
+                error!("Failed connecting to original_destination with {fail:#?}")
+            })?;
+
+        let mut request_sender = V::connect(tcp_stream, upgrade_tx)
+            .await
+            .inspect_err(|fail| error!("Handshake failed with {fail:#?}"))?;
+
+        V::send_request(&mut request_sender, request).await
+    }
+
+    /// Handles the incoming HTTP/V [`Request`].
+    ///
+    /// Checks if this [`Request`] contains an upgrade header, and if not, then checks if a header
+    /// matches one of the user specified filters.
+    ///
+    /// Helper function due to the fact that [`Service::call`] is not an `async` function.
+    pub(super) async fn handle_request(
+        request: Request<Incoming>,
+        original_destination: SocketAddr,
+        upgrade_tx: Option<oneshot::Sender<RawHyperConnection>>,
+        filters: Arc<DashMap<ClientId, Regex>>,
+        port: Port,
+        connection_id: ConnectionId,
+        request_id: RequestId,
+        matched_tx: Sender<HandlerHttpRequest>,
+    ) -> Result<Response<Full<Bytes>>, HttpTrafficError> {
+        if V::is_upgrade(&request) {
+            Self::unmatched_request(request, upgrade_tx, original_destination).await
+        } else if let Some(client_id) = header_matches(&request, &filters) {
+            let request = MatchedHttpRequest {
+                port,
+                connection_id,
+                client_id,
+                request_id,
+                request,
+            };
+
+            matched_request(request, matched_tx).await
+        } else {
+            Self::unmatched_request(request, None, original_destination).await
+        }
+    }
 }
